feat(core): move arc bounding box unit tests to the arcs package
