+++ packages/arcs/tests/.eslintrc.yml
@@ -1,0 +1,2 @@
+env:
+  jest: true
